% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{creditcards_data}
\alias{creditcards_data}
\title{Credit cards data request}
\usage{
creditcards_data(obs = 10)
}
\arguments{
\item{obs}{Number of rows returned}
}
\description{
Credit cards data request
}
