% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{custom_data}
\alias{custom_data}
\title{Custom table request}
\usage{
custom_data(obs = 10, data)
}
\arguments{
\item{obs}{Number of rows returned}

\item{data}{Data frame or tibble with column names and column types. First must can contain the names of the columns that we want to return. Second column must contain the value type. Accepted value types: "boolean","boolean_digit","buildingNumber","building_number","card_expiration","card_number","card_type","city","company_name","counter","country","countryCode","country_code","date","dateTime","date_time","ean","email","firstName","first_name","image","lastName","last_name","latitude","longText","long_text","longitude","name","null","number","phone","pokemon","postcode","state","streetAddress","streetName","street_address","street_name","text","upc","uuid","vat","website","word"}
}
\description{
Custom table request
}
