% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{fake_generate}
\alias{fake_generate}
\title{Generate fake data}
\usage{
fake_generate(x, ...)
}
\arguments{
\item{x}{an object of \code{fake} class.}
}
\value{
a list containing a fake dataset ($dataset) and
info of the variables ($info).
}
\description{
Generate fake data
}
