% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserve.R
\name{fake_preserve}
\alias{fake_preserve}
\title{Preserve columns}
\usage{
fake_preserve(x, ...)
}
\arguments{
\item{x}{an object of \code{fake} class.}

\item{...}{variables to be preserved. Preserved variables
won't be faked.}
}
\value{
a list containing a fake dataset ($dataset) and
info of the variables ($info).Ã
}
\description{
Preserve columns
}
