% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake.R
\name{fake_this}
\alias{fake_this}
\title{Fake a dataset}
\usage{
fake_this(x, ...)
}
\arguments{
\item{x}{a data.frame.}
}
\value{
a list containing a fake dataset ($dataset) and
info of the variables ($info).
}
\description{
Fake a dataset
}
