% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faker-package.R
\docType{package}
\name{faker-package}
\alias{faker}
\alias{faker-package}
\title{faker: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
