% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faker.R
\name{faker}
\alias{faker}
\title{Create a fake core request function}
\usage{
faker(resource, ..., quantity = 1, locale = "en_US", seed = NULL)
}
\arguments{
\item{resource}{resource}

\item{...}{other}

\item{quantity}{quantity}

\item{locale}{locale}

\item{seed}{seed}
}
\value{
JSON data
}
\description{
Create a fake core request function
}
\examples{
str(faker("images", width = 300))
}
