% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\name{faker_person}
\alias{faker_person}
\title{Get person}
\usage{
faker_person(
  gender = NULL,
  birthday_start = NULL,
  birthday_end = NULL,
  quantity = 1,
  locale = "en_US",
  seed = NULL
)
}
\arguments{
\item{gender}{gender}

\item{birthday_start}{birthday start}

\item{birthday_end}{birthday end}

\item{quantity}{quantity}

\item{locale}{locale}

\item{seed}{seed}
}
\value{
JSON data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
str(faker_person("male"))

}
