% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{images_data}
\alias{images_data}
\title{Images data request}
\usage{
images_data(
  obs = 10,
  image_type = "any",
  width = 640,
  height = 480,
  only_url_img = FALSE
)
}
\arguments{
\item{obs}{Number of rows returned}

\item{image_type}{Images category. Can be one of the following: any, animals, architecture, nature, people, tech, kittens, pokemon}

\item{width}{Width img in px. 640}

\item{height}{Height img in px. 480}

\item{only_url_img}{Boolean value. If TRUE only return url images vector}
}
\description{
Images data request
}
