% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{people_data}
\alias{people_data}
\title{People data request}
\usage{
people_data(
  obs = 10,
  gender = NULL,
  birthday_interval = c("1990-01-01", "2000-01-01")
)
}
\arguments{
\item{obs}{Number of rows returned. Max = 1000.}

\item{gender}{Boolean argument. "male" or "female" values accepted}

\item{birthday_interval}{Vector of two interval dates in yyyy-mm-dd format.}
}
\description{
People data request
}
