% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{products_data}
\alias{products_data}
\title{Producs data request}
\usage{
products_data(
  obs = 10,
  prices_interval = c(0, 100),
  tax = 30,
  category_type = "uuid"
)
}
\arguments{
\item{obs}{Number of rows returned}

\item{prices_interval}{Vector of prices. Must be 2 values: min and max}

\item{tax}{Tax percent. Value between 0 and 100}

\item{category_type}{Type of values returned on sku column. Values accepted: "integer", "string" and "uuid"}
}
\description{
Producs data request
}
