% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super.t.test.R
\name{super.t.test}
\alias{super.t.test}
\title{Super significant t.test}
\usage{
super.t.test(X, Y)
}
\arguments{
\item{X}{A distribution}

\item{Y}{Another distribution}
}
\value{
A significant t-test!
}
\description{
Always make sure your results are significant
}
\examples{
super.t.test(rnorm(50), rnorm(50))

}
\author{
Thomas Guillerme
}
