% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{texts_data}
\alias{texts_data}
\title{Texts data request}
\usage{
texts_data(obs = 10, characters = 120)
}
\arguments{
\item{obs}{Number of rows returned}

\item{characters}{String characters returned on each row}
}
\description{
Texts data request
}
