% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{users_data}
\alias{users_data}
\title{Users data request}
\usage{
users_data(obs = 10, genre = "female")
}
\arguments{
\item{obs}{Number of rows returned}

\item{genre}{Boolean argument. "male" or "female" values accepted}
}
\description{
Users data request
}
