% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faker.R
\name{faker_person_tidy}
\alias{faker_person_tidy}
\title{Get person data in tidy format}
\usage{
faker_person_tidy(
  gender = NULL,
  birthday_start = NULL,
  birthday_end = NULL,
  quantity = 1,
  locale = "en_US",
  seed = NULL
)
}
\arguments{
\item{gender}{gender}

\item{birthday_start}{birthday start}

\item{birthday_end}{birthday end}

\item{quantity}{quantity}

\item{locale}{locale}

\item{seed}{seed}
}
\value{
Tibble
}
\description{
Get person data in tidy format
}
\examples{
faker_person("male", quantity = 5)
}
