% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{faker_request}
\alias{faker_request}
\title{Create a fake core request function}
\usage{
faker_request(resource, ..., quantity = 1, locale = "en_US", seed = NULL)
}
\arguments{
\item{resource}{resource}

\item{...}{other}

\item{quantity}{quantity}

\item{locale}{locale}

\item{seed}{seed}
}
\value{
JSON data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
str(faker_request("images", width = 300))
}
