% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom ggplot theme}
\usage{
theme_custom(...)
}
\arguments{
\item{...}{Parameters to pass onto the base theme.}
}
\value{
ggplot theme
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
library("ggplot2")
mtcars2 <- within(mtcars, {
vs <- factor(vs, labels = c("V-shaped", "Straight"))
am <- factor(am, labels = c("Automatic", "Manual"))
cyl  <- factor(cyl)
gear <- factor(gear)
})

p1 <- ggplot(mtcars2) +
geom_point(aes(x = wt, y = mpg, colour = gear)) +
  labs(
    title = "Fuel economy declines as weight increases",
    subtitle = "(1973-74)",
    caption = "Data from the 1974 Motor Trend US magazine.",
    tag = "Figure 1",
    x = "Weight (1000 lbs)",
    y = "Fuel economy (mpg)",
    colour = "Gears"
  )

p1 + theme_custom()
}
}
\seealso{
Other visualization: 
\code{\link{default_theme}()}
}
\concept{visualization}
